+++ framework/db/oci/QueryBuilder.php
@@ -243,9 +243,11 @@ EOD;
      * @param string $table the table that new rows will be inserted into.
      * @param array $columns the column names
      * @param array $rows the rows to be batch inserted into the table
+     * @param boolean $ignore whether to excute insert ignore into, not support for oci
+     * @param boolean $replace whether to excute `repace into` instead of `insert into` , not support for oci
      * @return string the batch INSERT SQL statement
      */
-    public function batchInsert($table, $columns, $rows)
+    public function batchInsert($table, $columns, $rows, $ignore = false, $replace = false)
     {
         if (empty($rows)) {
             return '';